Creo la opción para realizzar el acta del partido

+++ src/EventListener/FotoEquipoUploadListener.php
@@ -44,12 +44,17 @@ class FotoEquipoUploadListener
     public function preUpdate(PreUpdateEventArgs $args)
     {
         $entity = $args->getEntity();
-        if($entity->getFoto()!==null && $entity->getFoto()!==''){
 
-            unlink(__DIR__.'/../../public/uploads/fotos/'.$entity->getFoto());
+        if($entity instanceof Equipo){
+            if($entity->getFoto()!==null && $entity->getFoto()!==''){
+
+                unlink(__DIR__.'/../../public/uploads/fotos/'.$entity->getFoto());
+            }
+            $this->uploadFile($entity);
         }
 
-        $this->uploadFile($entity);
+
+
     }
